Closed the loop with placeholder structs
The genetic algorithm is basically finished. But the struct that is
optimized is only a placeholder so far, containing a random initialized
number that gets added in genetic combination.

The next step will be to read in the image, draw triangles and calculate
the correct fitness.
Afterwards some docstrings and tests will be necessary.